Big conversion to iFlow3.0, with new underlying structure and new package structure. Example files have been update accordingly. This version underwent limited testing, so should be regarded as beta-testing version

+++ src/util/diagnostics/LogConfigurator.py
@@ -2,11 +2,14 @@
 Log Configurator
 Creates a top-level logger and contains methods for configuring the logger
 
-Date: 28-04-15
+
+Original date: 28-04-15
+Updated: 03-02-22
 Authors: Y.M. Dijkstra
 """
 import logging
 import os
+from src.config import ConsoleLoggingLevel, FileLoggingLevel
 from src.util.diagnostics import DiagFormatter
 
 
@@ -21,21 +24,35 @@ class LogConfigurator:
         Parameters:
             name - name of root level logger. Equals the name of the top level package (e.g. src)
         """
-        self.logger = logging.getLogger(name)
-        self.logger.setLevel(logging.INFO)
-        return
+        self.logger = logging.root   # handle to root logger
 
-    def getCount(self):
-        return self.logger.handlers[0].level2count
+        self.ConsoleLoggingLevelUp = ConsoleLoggingLevel.upper()
+        self.FileLoggingLevelUp = FileLoggingLevel.upper()
+
+        if not self.ConsoleLoggingLevelUp in ['DEBUG', 'INFO', 'WARNING', 'ERROR', 'CRITICAL', 'OFF']:
+            logging.info('Incorrect value for variable LoggingLevel in the iFlow config file. Please set to DEBUG, INFO, WARNING, ERROR, CRITICAL (see python logging system) or OFF. Logging level now set to INFO.')
+            self.ConsoleLoggingLevelUp = 'INFO'
+        if not self.FileLoggingLevelUp in ['DEBUG', 'INFO', 'WARNING', 'ERROR', 'CRITICAL', 'OFF']:
+            logging.info('Incorrect value for variable LoggingLevel in the iFlow config file. Please set to DEBUG, INFO, WARNING, ERROR, CRITICAL (see python logging system) or OFF. Logging level now set to INFO.')
+            self.FileLoggingLevelUp = 'INFO'
+
+        self.__setRootLevel(self.ConsoleLoggingLevelUp, self.FileLoggingLevelUp)
+        for i in self.logger.handlers:
+            self.logger.removeHandler(i)
+        return
 
     def makeConsoleLog(self):
         """Configure the format for the console logger. Makes a handler for messages on the console
         """
-        console = logging.StreamHandler() #MsgCounterHandler()
-        console.setLevel(logging.INFO)
-        consoleFormat   = DiagFormatter('console')
-        console.setFormatter(consoleFormat)
-        if not any([isinstance(i, logging.StreamHandler) for i in self.logger.handlers]):       # add a streamHandler logger if one does not exist
+        if not self.ConsoleLoggingLevelUp == 'OFF':
+            console = logging.StreamHandler()
+            self.__setLogLevel(self.ConsoleLoggingLevelUp, console)
+
+            consoleFormat   = DiagFormatter('console')
+            console.setFormatter(consoleFormat)         # assign formatter to handler
+
+
+            # if not any([isinstance(i, logging.StreamHandler) for i in self.logger.handlers]):       # add a streamHandler logger if one does not exist
             self.logger.addHandler(console)
         return
 
@@ -45,39 +62,57 @@ class LogConfigurator:
         Parameters:
             filepath - path to the input file
         """
-        # extract path and extension of input file 
-        filepath.replace('\\','/')
-        length = filepath.rfind('/')            # finds index of last occurrence of '/' or returns -1
-        logFileLocation = filepath[:length+1]
-        
-        logFileExtension = filepath[-3:]
-        logFilePath = (logFileLocation+'diag.'+logFileExtension)
+        # extract path and extension of input file
+        if not self.FileLoggingLevelUp == 'OFF':
+            filepath.replace('\\','/')
+            length = filepath.rfind('/')            # finds index of last occurrence of '/' or returns -1
+            logFileLocation = filepath[:length+1]
+            logFileName = filepath[length+1:-4]
+            if logFileLocation[-1] != '/':
+                logFileLocation = logFileLocation+'/'
 
-        # clear file if needed
-        try:
-            os.remove(logFilePath)
-        except:
-            pass
+            logFileExtension = filepath[-3:]
+            logFilePath = (logFileLocation+'diag_'+logFileName + '.'+logFileExtension)
 
-        # format and add handlers
-        fileHandler = logging.FileHandler(logFilePath)
-        fileHandler.setLevel(logging.INFO)
-        fileHandlerFormat = DiagFormatter('file')
-        fileHandler.setFormatter(fileHandlerFormat)
-        self.logger.addHandler(fileHandler)
-        
+            # clear file if needed
+            try:
+                os.remove(logFilePath)
+            except:
+                pass
+
+            # format and add handlers
+            fileHandler = logging.FileHandler(logFilePath)
+            self.__setLogLevel(self.FileLoggingLevelUp, fileHandler)
+
+            fileHandlerFormat = DiagFormatter('file')
+            fileHandler.setFormatter(fileHandlerFormat)
+
+            self.logger.addHandler(fileHandler)
         return
 
-# class MsgCounterHandler(logging.StreamHandler):
-#     level2count = None
-#
-#     def __init__(self, *args, **kwargs):
-#         super(MsgCounterHandler, self).__init__(*args, **kwargs)
-#         self.level2count = {}
-#
-#     def emit(self, record):
-#         super(MsgCounterHandler, self).emit(record)
-#         l = record.levelname
-#         if (l not in self.level2count):
-#             self.level2count[l] = 0
-#         self.level2count[l] += 1
+    def __setLogLevel(self, level, handler):
+        if level == 'DEBUG':
+            handler.setLevel(logging.DEBUG)
+        elif level == 'INFO':
+            handler.setLevel(logging.INFO)
+        elif level == 'WARNING':
+            handler.setLevel(logging.WARNING)
+        elif level == 'ERROR':
+            handler.setLevel(logging.ERROR)
+        elif level == 'CRITICAL':
+            handler.setLevel(logging.CRITICAL)
+        return
+
+    def __setRootLevel(self, level1, level2):
+        if level1 == 'DEBUG' or level2 == 'DEBUG':
+            logging.basicConfig(level=logging.DEBUG)
+        elif level1 == 'INFO' or level2 == 'INFO':
+            logging.basicConfig(level=logging.INFO)
+        elif level1 == 'WARNING' or level2 == 'WARNING':
+            logging.basicConfig(level=logging.WARNING)
+        elif level1 == 'ERROR' or level2 == 'ERROR':
+            logging.basicConfig(level=logging.ERROR)
+        elif level1 == 'CRITICAL' or level2 == 'CRITICAL':
+            logging.basicConfig(level=logging.CRITICAL)
+        return
+
